backend/requirements.txt
fastapi
uvicorn[standard]
beanie
motor
python-dotenv
pydantic[email]
pydantic-settings